+++ pages/src/types/RedirectsCollectionConfigAttributes.ts
@@ -1,0 +1,5 @@
+/** The incoming attributes for the redirects collection config. */
+export type IncomingRedirectsCollectionConfigAttributes = {}
+
+/** The attributes for the redirects collection config after they have been processed using the incoming config attributes. */
+export type RedirectsCollectionConfigAttributes = {}